Conflicts:
	fusinvdeploy/inc/config.class.php
	fusinvdeploy/inc/package.class.php
	fusinvdeploy/install/update.php
	fusinvdeploy/setup.php
	fusinvinventory/inc/antivirus.class.php
	fusinvinventory/inc/libhook.class.php
	fusinvsnmp/inc/communicationnetdiscovery.class.php
	fusinvsnmp/inc/communicationsnmpquery.class.php
	fusinvsnmp/inc/networkport.class.php
	fusinvsnmp/inc/networkportlog.class.php
	fusioninventory/hook.php
	fusioninventory/inc/agent.class.php
	fusioninventory/inc/communication.class.php
	fusioninventory/install/update.php
	fusioninventory/phpunit/AllTests.php
	fusioninventory/phpunit/FusinvInstall/AllTests.php
	fusioninventory/phpunit/InventoryComputer/AllTests.php
	fusioninventory/phpunit/Netdiscovery/AllTests.php
	fusioninventory/phpunit/Netinventory/Hub/AllTests.php
	fusioninventory/phpunit/Rules/AllTests.php
	fusioninventory/tools/phpunit/InventorySNMP.php

+++ fusinvinventory/inc/import_monitor.class.php
@@ -169,6 +169,14 @@ class PluginFusinvinventoryImport_Monitor extends CommonDBTM {
       if ($_SESSION["plugin_fusinvinventory_no_history_add"]) {
          $array['_no_history'] = $_SESSION["plugin_fusinvinventory_no_history_add"];
       }
+      if ($type == "add") { // Case where have same monitor xx times
+         $a_find = $Computer_Item->find("`computers_id`='".$items_id."'
+            AND `itemtype` = 'Monitor'
+            AND `items_id`='".$monitor_id."'");
+         if (count($a_find) > 0) {
+            return;
+         }
+      }
       $devID = $Computer_Item->add($array);
       return $devID;
    }
@@ -189,15 +197,16 @@ class PluginFusinvinventoryImport_Monitor extends CommonDBTM {
       if ($pfConfig->getValue($_SESSION["plugin_fusinvinventory_moduleid"],
                  "import_monitor") != '0') {
          $Computer_Item = new Computer_Item();
-         $Computer_Item->getFromDB($items_id);
-         if ($Computer_Item->fields['computers_id'] == $idmachine) {
-            $input = array();
-            $input['id'] = $items_id;
-            $input['itemtype'] = 'Monitor';
-            if ($_SESSION["plugin_fusinvinventory_no_history_add"]) {
-               $input['_no_history'] = $_SESSION["plugin_fusinvinventory_no_history_add"];
+         if ($Computer_Item->getFromDB($items_id)) {
+            if ($Computer_Item->fields['computers_id'] == $idmachine) {
+               $input = array();
+               $input['id'] = $items_id;
+               $input['itemtype'] = 'Monitor';
+               if ($_SESSION["plugin_fusinvinventory_no_history_add"]) {
+                  $input['_no_history'] = $_SESSION["plugin_fusinvinventory_no_history_add"];
+               }
+               $Computer_Item->delete($input, 0, $_SESSION["plugin_fusinvinventory_history_add"]);
             }
-            $Computer_Item->delete($input, 0, $_SESSION["plugin_fusinvinventory_history_add"]);
          }
       }
    }
